Adjusted max_jobs
i adjusted max_jobs because i kept getting a seg fault and i thought it was because it was passing an invalid number. I also thought that there was incorrect memory access so I attempted to change it all to a fixed job array size but now Im getting a "Cannot access memory at address".